# Remove ip16 type, change it into uint128 type, use it in IP also