The guts of this change are:

    // uint128 represents a uint128 using two uint64s.
    // Index 0 contains the high bits, index 1 contains the low.
    type uint128 [2]uint64

Then it replaces IP.hi and IP.lo with IP.addr of type uint128. Same
shape in memory, but with a name and type and lets us pass around both
parts together in one variable (as now used by IPRange.Prefix and
IPRange.Prefixes), including putting methods on them (moved from
the now-deleted ip16 type)

No change in inlining.