@bradfitz 2020-12-30
Feel free to merge if it's blocking you playing. Otherwise I'll merge tomorrow giving others time to comment too if they wish.