@AlexanderYastrebov 2020-12-31
> No change in inlining.
```
$ git checkout 318330f177ab307858b73763bf7e8715db66a4a5~1 && go test -run=NONE -bench=BenchmarkParseIP > /tmp/old && git checkout 318330f177ab307858b73763bf7e8715db66a4a5 && go test -run=NONE -bench=BenchmarkParseIP > /tmp/new && benchcmp -changed /tmp/old /tmp/new
HEAD is now at 4ef2ea6 Add a few more alloc-free methods.
Previous HEAD position was 4ef2ea6 Add a few more alloc-free methods.
HEAD is now at 318330f Remove ip16 type, change it into uint128 type, use it in IP also
benchmark                              old ns/op     new ns/op     delta
BenchmarkParseIP/v4-8                  23.0          28.9          +25.65%
BenchmarkParseIP/v6-8                  77.8          92.7          +19.15%
BenchmarkParseIP/v6_ellipsis-8         56.3          70.1          +24.51%
BenchmarkParseIP/v6_v4-8               64.0          87.5          +36.72%
BenchmarkParseIP/v6_zone-8             145           163           +12.41%
BenchmarkParseIPPort/v4-8              42.8          54.2          +26.64%
BenchmarkParseIPPort/v6-8              99.5          127           +27.64%
BenchmarkParseIPPort/v6_ellipsis-8     79.9          98.7          +23.53%
BenchmarkParseIPPort/v6_v4-8           88.6          111           +25.28%
BenchmarkParseIPPort/v6_zone-8         204           197           -3.43%
```